requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
apscheduler
python-dotenv
langchain-google-genai
langchain-core
twilio
